Revised #42 Organized controller into class to call functions in index file.
