+++ controller/homeContr.class.php
@@ -1,0 +1,7 @@
+<?php
+
+Class HomeContr {
+    public static function home(){
+        include './view/homeView.php';
+    }
+}
