--- controller/landing.php
@@ -1,3 +1,0 @@
-<?php
-
-include 'view/landingView.php';
