+++ index.php
@@ -1,39 +1,42 @@
 <?php
     session_start();
 
-    include_once "controller/_paths.php";
+    include_once "./controller/_paths.php";
+    include_once "./controller/previewContr.class.php";
+    include_once "./controller/homeContr.class.php";
+    include_once "./controller/adminContr.class.php";
+    include_once "./controller/usersContr.class.php";
+    include_once "./controller/challengeContr.class.php";
 
     $action = $_GET['action'] ?? '';
 
-
-
     switch($action) {
         case 'admin':
-            include 'controller/admin.php';
+            AdminContr::admin();
             break;
         case 'admin_edit': 
-            include 'controller/admin_edit.php'; 
+            AdminContr::adminEdit();
             break;
         case 'register':
-            include 'controller/register.php'; 
+            UsersContr::registerUser();
             break;
         case 'create-challenge':
-            include './controller/create-challenge.php';
+            ChallengeContr::createChallenge();
             break;
         case 'list-challenges':
-            include './controller/list-challenges.php';
+            ChallengeContr::listChallenges();
             break;
         case 'challenge-specific':
-            include './controller/challenge-specific.php';
-            break;
-        case "preview":
-            include 'controller/landing.php';
+            ChallengeContr::showChallengeInfo();
             break;
         case 'login':
-            include './controller/loginController.php';
+            UsersContr::loginUser();
+            break;
+        case 'home':
+            HomeContr::home();
             break;
         default:
-            include './controller/display-home.php';
+            PreviewContr::preview();
             break;
     }
 ?>
