+++ view/header.php
@@ -6,13 +6,13 @@
 <header>
     <nav id="menu">
         <ul>
-            <li><a href="<?= $root ?>" style="<?= $action == '' ? $highlight:'' ?>">On My Way</a></li>
-            <li><a href="<?= $createChallenge_path ?>" style="<?= $action == 'create-challenge' ? $highlight:'' ?>">Add challenge</a></li>
+            <li><a href="<?= ROOT ?>" style="<?= $action == '' ? $highlight:'' ?>">On My Way</a></li>
+            <li><a href="<?= CREATE_CHALLENGE_PATH ?>" style="<?= $action == 'create-challenge' ? $highlight:'' ?>">Add challenge</a></li>
             <!-- if loggedin then show this -->
-                <li><a href="<?= $$userProfile_path ?>" style="<?= $action == 'my-profile' ? $highlight:'' ?>">See My Profile</a></li>
+                <li><a href="<?= USER_PROFILE_PATH ?>" style="<?= $action == 'my-profile' ? $highlight:'' ?>">See My Profile</a></li>
             <!-- end if -->
             <!-- if the user is admin == true, let's delete this link later and just send the admin directly to admin page -->
-                <li><a href="<?= $admin_path ?>" style="<?= $action == 'admin' ? $highlight:'' ?>">Admin</a></li>
+                <li><a href="<?= ADMIN_PATH ?>" style="<?= $action == 'admin' ? $highlight:'' ?>">Admin</a></li>
             <!-- end if -->
         </ul>
     </nav>
