+++ view/homeView.php
