+++ view/previewView.php
@@ -4,8 +4,8 @@ ob_start();
 ?>
 
 <h1>Welcome to On My Way</h1>
-<h2><a href="../on-my-way/index.php?action=register">Join Now</a></h2>
-<h3>Please click here to <a href="<?= $login_path ?>">log in</a> / <a href="<?= $register_path ?>">register</a></h3>
+<h2><a href="<?= REGISTER_PATH ?>">Join Now</a></h2>
+<h3>Please click here to <a href="<?= LOGIN_PATH ?>">log in</a> / <a href="<?= REGISTER_PATH ?>">register</a></h3>
 <div id="map"></div>
